Fixes to drop action and css to free layout

+++ src/action/drop-action.ts
@@ -1,9 +1,11 @@
 import { Vine } from '@grapevine';
 import { scanArray } from '@gs-tools/rxjs';
 import { Observable } from '@rxjs';
-import { switchMap, tap, withLatestFrom } from '@rxjs/operators';
+import { map, switchMap, tap, withLatestFrom } from '@rxjs/operators';
+
 import { BaseAction } from '../core/base-action';
 import { TriggerType } from '../core/trigger-spec';
+
 import { $pickService } from './pick-service';
 
 export class DropAction extends BaseAction {
@@ -12,21 +14,30 @@ export class DropAction extends BaseAction {
   }
 
   protected onTrigger(vine: Vine): Observable<unknown> {
+    const components$ = $pickService.get(vine).pipe(
+        switchMap(pickService => {
+          return pickService.getComponents()
+              .pipe(
+                  scanArray(),
+                  map(components => ({components, pickService})),
+              );
+        }),
+    );
+
     return this.parentNode$
         .pipe(
-            withLatestFrom($pickService.get(vine)),
-            switchMap(([parentNode, pickService]) => pickService.getComponents()
-                .pipe(
-                    scanArray(),
-                    tap(components => {
-                      const nextComponent = components[components.length - 1] || null;
-                      pickService.deleteAt(components.length - 1);
+            withLatestFrom(components$),
+            tap(([parentNode, {components, pickService}]) => {
+              const nextComponent = components[components.length - 1] || null;
+              if (!nextComponent) {
+                return null;
+              }
 
-                      parentNode.appendChild(nextComponent);
-                      return nextComponent;
-                    }),
-                ),
-            ),
+              pickService.deleteAt(components.length - 1);
+
+              parentNode.appendChild(nextComponent);
+              return nextComponent;
+            }),
         );
   }
 }
